added individual board components rendering with dummy data
